Reorganize server start
