+++ packages/api/index.js
@@ -1,87 +1,7 @@
 require('dotenv').config()
 
-const Dash = require('dash')
-const Fastify = require('fastify')
-const cors = require('@fastify/cors')
-const Routes = require('./src/routes')
+const server = require('./src/server')
 
-const ServiceNotAvailableError = require("./src/errors/ServiceNotAvailableError");
-const MainController = require("./src/controllers/MainController");
-const TransactionsController = require("./src/controllers/TransactionsController");
-const BlocksController = require("./src/controllers/BlocksController");
-const DocumentsController = require("./src/controllers/DocumentsController");
-const IdentitiesController = require("./src/controllers/IdentitiesController");
-const packageVersion = require('./package.json').version
-const Worker = require('./src/worker/index')
-const {BLOCK_TIME} = require("./src/constants");
-const DataContractsController = require("./src/controllers/DataContractsController");
-
-function errorHandler(err, req, reply) {
-    if (err instanceof ServiceNotAvailableError) {
-        return reply.status(403).send({error: 'tenderdash backend is not available'})
-    }
-
-    if (err?.constructor?.name === 'InvalidStateTransitionError') {
-        const [error] = err.getErrors()
-        const {code, message} = error
-
-        return reply.status(500).send({error: message, code})
-    }
-
-    console.error(err)
-    reply.status(500)
-
-    reply.send({error: err.message})
-}
-
-let status;
-
-const init = async () => {
-    const client = new Dash.Client()
-    await client.platform.initialize()
-
-    const worker = new Worker()
-
-    worker.setHandler(async () => {
-        try {
-            status = true
-        } catch (e) {
-        }
-    })
-    worker.start(BLOCK_TIME)
-
-    const fastify = Fastify()
-
-    await fastify.register(cors, {
-        // put your options here
-    })
-
-    const knex = require('knex')({
-        client: 'pg',
-        connection: {
-            host: process.env["POSTGRES_HOST"],
-            port: process.env["POSTGRES_PORT"],
-            user: process.env["POSTGRES_USER"],
-            database: process.env["POSTGRES_NAME"],
-            password: process.env["POSTGRES_PASS"],
-            ssl: process.env["POSTGRES_SSL"] ? { rejectUnauthorized: false } : false,
-        }
-    });
-
-    await knex.raw('select 1+1');
-
-    const mainController = new MainController(knex)
-    const blocksController = new BlocksController(knex)
-    const transactionsController = new TransactionsController(client, knex)
-    const dataContractsController = new DataContractsController(knex)
-    const documentsController = new DocumentsController(knex)
-    const identitiesController = new IdentitiesController(knex)
-
-    Routes({fastify, mainController, blocksController, transactionsController, dataContractsController, documentsController, identitiesController})
-
-    fastify.setErrorHandler(errorHandler)
-    fastify.listen({ host: "0.0.0.0", port: 3005, listenTextResolver: (address) => console.log(`Platform indexer API has started on the ${address}`)});
-}
-
-
-init().then((`Platform Explorer backend v${packageVersion}`))
+server.start()
+    .then(server.listen)
+    .then(() => console.log(`Platform Explorer API started`))
